Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
bcrypt
PyJWT
psycopg2-binary